Create Assignment 2
This Python program calculates the shipping cost for sending a certain number of packages. It first asks the user to enter how many packages they want to send, and then it asks whether the shipping type is standard (“r”) or express (“e”). Based on the user’s choice, the program assigns a rate of $10 for standard shipping or $15 for express shipping; if the user enters anything else, it marks the shipping type as invalid and sets the rate to $0. It then multiplies the number of packages by the rate to determine the total delivery cost. Finally, the program displays the selected shipping type and the number of packages entered.